Squashed commit of the following:
commit 156135e543fbb978c7f358aa50c445e88227e226
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 18:02:06 2022 -0400

    reverted change

commit 5cb4f848a25acb920878edd1ae1e202fd4c31ea0
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 18:01:30 2022 -0400

    removed auxiliary changes

commit fc129cef26be4a54ffe8964384313aa907a5a800
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:59:30 2022 -0400

    updated to include match MapSchema

commit 24cf71755bb61c4243729857e6659f16b3d2fb27
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:58:19 2022 -0400

    simplified exports

commit ea383d05aeaf4d2981596036cf48e543ea28308b
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:56:35 2022 -0400

    added new line at end of file

commit e23ceaeb48a4189605c3261d89556f872bd0cc9e
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:56:11 2022 -0400

    reverted change

commit 93d99af66c0ab7d045153c294330b4a8c6c422d5
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:54:25 2022 -0400

    updated exports to fix tests

commit 58a7aa11d625ce72cf547870e800f1050a96118c
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:51:29 2022 -0400

    reverted changes

commit 6333d602f9a41cf79a911a36a05a4dd0250884fe
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:50:19 2022 -0400

    removed unused type

commit b11f201e58925f443432dd25e42cf4308a581a48
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:49:59 2022 -0400

    fixed typing conflicting with function

commit c9d6e1ae1694951c28e24606562437eaa7f275fe
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:48:57 2022 -0400

    reverted changes

commit fc84e377980d3568902910ef4378d0c5a1b74790
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:48:27 2022 -0400

    moved types around to better reflect domains

commit 7c0fc9f72cc74569a01e6b00528beb30c4bb9f4c
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 16:40:00 2022 -0400

    refactored to reduce file bloat and impact on other files

commit ac43f0f81df80a67c7232e5efdc3aaed68692925
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 15:21:49 2022 -0400

    refactored to use optional chaining

commit e0df843471d5b402f5a111162a3168c59e751480
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 14:51:53 2022 -0400

    split annotations file into separate files by type

commit e3e69c0455fdb3ac7a88611ae77aa4cc241275c5
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Thu Mar 31 14:21:35 2022 -0400

    simplified types and implementation

commit 8dfe1f09b19c10454e257e454b2b5d4058443e9f
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Tue Mar 29 20:40:55 2022 -0400

    changed to use Schema.is

commit c7f0300a49e354b5b01b9511f7bfa877b6597997
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Tue Mar 29 20:39:38 2022 -0400

    reverted "fix"

commit 4747b7315fe77aa46f7ff9f99f054955f5fb1cb9
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Tue Mar 29 20:34:52 2022 -0400

    fixed isPrototypeOf

commit d723bbc9e752215ace88206bd65a4272fcd1c3e6
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Tue Mar 29 20:15:29 2022 -0400

    simplified implementation

commit cd7847bde2b3ad0451d50117105a7371e3742e1b
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Fri Mar 18 03:00:30 2022 -0400

    added defaults for generics

commit 8c0c52ae99930d044d44aed1dd56edf66c4e38f2
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Fri Mar 18 02:53:15 2022 -0400

    updated exports

commit 9f06c015b57b35f4cb04ba69c37f0f87ec7331ed
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Fri Mar 18 02:53:00 2022 -0400

    refactored to address file length issue

commit 706fb8e3819267108fca484af071e4037e92e82e
Author: Wesley Clements <wesley.clements@gmail.com>
Date:   Fri Mar 18 02:00:31 2022 -0400

    fixed strict equality for null
